iptables-nftables-compat: link compat manpages to the originals
There are no specific manpages for the -compat tools right now.

Writting such manpages is something I would like to discuss with upstream.
I think the best sollution by now is to simply redirect users to the originals.
For this, a dependency with the original tools is required.

It should be noted that such a dependency is not weird, since its very likely
that users of the iptables-nftables-compat package are triying to migrate
their ruleset, so they need the original tools anyway.

Signed-off-by: Arturo Borrero Gonzalez <arturo.borrero.glez@gmail.com>