libxtables12: fix Replaces and Breaks relationships
The problem with this package and libxtables11 not being co-installable
was introduced in the first snapshot version which did not get the
SONAME bumped for the package name. But older versions should still be
co-installable as they have different SONAMES. Change the direction
of the verioned dependencies to cover this.

In addition using a >= should be safe here as long as a libxtables11
is not reintroduced.

Closes: #845800
Signed-off-by: Guillem Jover <guillem@debian.org>
Signed-off-by: Arturo Borrero Gonzalez <arturo@debian.org>